+++ workspace/apps/server/src/controllers/account.ts
@@ -401,7 +401,7 @@ export const updateInfoMutation = procedure //
                     options.ctx.S.log.error(error);
                     return {
                         code: -1,
-                        message: error,
+                        message: String(error),
                         data: null,
                     };
             }